pyproject.toml
[tool.poetry]
name = "gerd"
version = "0.2.0"
description = ""
authors = ["cto"]
license = ""
readme = "README.md"

[tool.poetry.scripts]
start = "main:app"

[tool.poetry.dependencies]
torch = [
    { version = ">=2.0.0, !=2.0.1", platform = "!=darwin", source = "torch_cpu" },
    { version = ">=2.0.0, !=2.0.1", platform = "darwin", source = "PyPI" },
]
orjson = "=3.10.5"
python = ">=3.12, <4.0"
transformers = "^4.41.2"
faiss-cpu = { version = "~1.7.4", optional = true }
faiss-gpu = { version = "~1.7.2", optional = true }
ipykernel = "^6.29.4"
langchain = "^0.2.6"
pypdf = "^4.2.0"
sentence-transformers = "^3.0.1"
uvicorn = "^0.30.1"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.4"
poethepoet = "^0.26.1"
fastapi = "^0.111.0"
requests = "^2.32.3"
jinja2 = "^3.1.4"
langchain-community = "^0.2.6"
llama-cpp-python = "^0.2.81"
langchain-huggingface = "^0.0.3"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
pytest = "^8.2.2"
pre-commit = "^3.7.1"
types-pyyaml = "^6.0.12.20240311"
types-requests = "^2.32.0.20240622"
pytest-mock = "^3.14.0"

[tool.poetry.extras]
cpu = ["faiss-cpu"]
gpu = ["faiss-gpu"]

[[tool.poetry.source]]
name = "torch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poe.tasks]
test = "poetry run pytest -x"
qa = "poetry run gradio team_red/frontends/qa_frontend.py"
gen = "poetry run gradio team_red/frontends/gen_frontend.py"
server = "poetry run python -m gerd.backends.rest_server"

[tool.ruff.lint]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # isort (imports) styling
    "N",    # pep8 naming checks
    "B",    # bugbear checks
    "ANN",  # type annotations
    "S",    # bandid code security
    "C",    # improve code comprehension
    "EM",   # use recommended error message composition
    "G",    # use recommended logging format
    "T20",  # no print allowed
    "PT",   # pytest styling recommendations
    "SLF",  # prevent access to private members
    "SIM",  # code simplification recommendations
    "TCH",  # checks for type checking imports
    "ARG",  # no unused arguments
    "PERF", # performance anti pattern
    # "FURB",  # enforces modern python
    # "D",     # enforce documented functions [will be enabled soon]
]

# 2. Avoid enforcing line-length violations (`E501`)
ignore = [
    "ANN101", # do not enforce self type annotation
    "ANN102", # do not enforce cls type annoation
]

# 3. Fix everything except flake8-bugbear (`B`) violations.
fixable = ["ALL"]
unfixable = ["B"]

# 4. Things to ignore
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"] # assert is fine in pytest
# "__init__.py" = ["E402"]
# 
# "**/{tests,docs,tools}/*" = ["E402"]

# [tool.ruff.format]
# # 5. Use single quotes for non-triple-quoted strings.
# quote-style = "single"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
